--- a/vllm/v1/worker/gpu_worker.py
+++ b/vllm/v1/worker/gpu_worker.py
@@ -164,6 +164,9 @@ class Worker(WorkerBase):
             # This env var set by Ray causes exceptions with graph building.
             os.environ.pop("NCCL_ASYNC_ERROR_HANDLING", None)
             self.device = torch.device(f"cuda:{self.local_rank}")
+            if not envs.VLLM_USE_DEFAULT_STREAM:
+                self.cuda_stream = torch.cuda.Stream(device=self.device)
+                torch.cuda.set_stream(self.cuda_stream)
             current_platform.set_device(self.device)
 
             _check_if_gpu_supports_dtype(self.model_config.dtype)