## Current Issue
Currently, our gpu model runners runs in the default CUDA stream. This works fine in a non-disagg environment with no multiple CUDA stream running together. However, in a disagg environment, other than the normal model forward execution on GPU, there would be other KV saving/loading operations (mostly H2D, D2H) running in their own CUDA streams to avoid interleaving with the model forward execution. With gpu model runner runs in default CUDA stream, the blocking copy operations such like `valid_sampled_token_ids = sampled_token_ids.tolist()` would block other copy operations scheduled to run in other CUDA streams even when copy engine is idle.

<img width="1694" height="798" alt="Screenshot 2025-07-22 at 2 54 08 PM" src="https://github.com/user-attachments/assets/f5adcb89-416c-431f-8c3b-d046b78b8058" />
<img width="1690" height="850" alt="Screenshot 2025-07-22 at 2 54 18 PM" src="https://github.com/user-attachments/assets/9188b9c6-0bc8-46b9-8cd5-3e8473061435" />
<img width="1678" height="798" alt="Screenshot 2025-07-22 at 2 54 23 PM" src="https://github.com/user-attachments/assets/69d93b35-936d-45ff-ae1b-a62457ff0dc3" />


## Change
Let GPU model runner runs in its own CUDA stream, other than the default CUDA stream.

## Test for Non Disagg
Bring up vLLM server
```
VLLM_USE_V1=1  vllm serve $llama4 --disable-log-requests -tp 8 --max-num-seqs 64 --no-enable-prefix-caching --max_num_batched_tokens=8000
```

Load test for non-disagg setup and observe no perf difference:

Before
```
Maximum request concurrency: 32
100%|██████████| 640/640 [01:04<00:00,  9.85it/s]
============ Serving Benchmark Result ============
Successful requests:                     640       
Maximum request concurrency:             32        
Benchmark duration (s):                  64.96     
Total input tokens:                      1278506   
Total generated tokens:                  94259     
Request throughput (req/s):              9.85      
Output token throughput (tok/s):         1450.97   
Total Token throughput (tok/s):          21131.53  
---------------Time to First Token----------------
Mean TTFT (ms):                          299.10    
Median TTFT (ms):                        258.22    
P99 TTFT (ms):                           1264.80   
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          19.95     
Median TPOT (ms):                        20.21     
P99 TPOT (ms):                           23.11     
---------------Inter-token Latency----------------
Mean ITL (ms):                           19.94     
Median ITL (ms):                         13.26     
P99 ITL (ms):                            151.18    
==================================================

Maximum request concurrency: 64
100%|██████████| 640/640 [00:48<00:00, 13.12it/s]
============ Serving Benchmark Result ============
Successful requests:                     640       
Maximum request concurrency:             64        
Benchmark duration (s):                  48.80     
Total input tokens:                      1278506   
Total generated tokens:                  94382     
Request throughput (req/s):              13.12     
Output token throughput (tok/s):         1934.11   
Total Token throughput (tok/s):          28133.74  
---------------Time to First Token----------------
Mean TTFT (ms):                          358.76    
Median TTFT (ms):                        254.04    
P99 TTFT (ms):                           2373.02   
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          30.59     
Median TPOT (ms):                        31.40     
P99 TPOT (ms):                           36.43     
---------------Inter-token Latency----------------
Mean ITL (ms):                           30.56     
Median ITL (ms):                         16.21     
P99 ITL (ms):                            152.50  
``` 

After
```
Maximum request concurrency: 32
100%|██████████| 640/640 [01:06<00:00,  9.67it/s]
============ Serving Benchmark Result ============
Successful requests:                     640       
Maximum request concurrency:             32        
Benchmark duration (s):                  66.17     
Total input tokens:                      1278506   
Total generated tokens:                  94444     
Request throughput (req/s):              9.67      
Output token throughput (tok/s):         1427.23   
Total Token throughput (tok/s):          20747.98  
---------------Time to First Token----------------
Mean TTFT (ms):                          272.46    
Median TTFT (ms):                        250.93    
P99 TTFT (ms):                           1100.20   
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          20.52     
Median TPOT (ms):                        20.85     
P99 TPOT (ms):                           23.75     
---------------Inter-token Latency----------------
Mean ITL (ms):                           20.51     
Median ITL (ms):                         13.43     
P99 ITL (ms):                            151.92    
==================================================

Maximum request concurrency: 64
100%|██████████| 640/640 [00:49<00:00, 13.02it/s]
============ Serving Benchmark Result ============
Successful requests:                     640       
Maximum request concurrency:             64        
Benchmark duration (s):                  49.16     
Total input tokens:                      1278506   
Total generated tokens:                  94680     
Request throughput (req/s):              13.02     
Output token throughput (tok/s):         1926.04   
Total Token throughput (tok/s):          27934.17  
---------------Time to First Token----------------
Mean TTFT (ms):                          390.31    
Median TTFT (ms):                        284.53    
P99 TTFT (ms):                           2366.98   
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          30.53     
Median TPOT (ms):                        31.55     
P99 TPOT (ms):                           35.03     
---------------Inter-token Latency----------------
Mean ITL (ms):                           30.51     
Median ITL (ms):                         16.37     
P99 ITL (ms):                            154.25    
==================================================
```



## Test for Disagg
Runs with a vendor internal disagg implementation (kv connector interface based), and observe TTIT wins.

Before
```
Ran 1029/1029 requests in 133.00s
Success rate:        100.00%
QPS:                 7.74
Avg latency:         4.841s
Avg TTFT (client):   137.37ms
P50 TTFT (client):   130.12ms
P99 TTFT (client):   364.42ms
Avg TTIT (client):   31.35ms
P50 TTIT (client):   31.12ms
P99 TTIT (client):   38.27ms
Avg TTFT (server):   134.39ms
Avg TTIT (server):   31.29ms
Avg prefill len:     2199.72 tokens
P50 prefill len:     2199.00 tokens
P99 prefill len:     2231.00 tokens
Avg decode len:      150.00 tokens
P50 decode len:      150.00 tokens
P99 decode len:      150.00 tokens
```
After
```
Ran 1026/1026 requests in 132.34s
Success rate:        100.00%
QPS:                 7.75
Avg latency:         4.550s
Avg TTFT (client):   128.50ms
P50 TTFT (client):   125.91ms
P99 TTFT (client):   205.68ms
Avg TTIT (client):   29.47ms
P50 TTIT (client):   29.59ms
P99 TTIT (client):   30.00ms
Avg TTFT (server):   124.88ms
Avg TTIT (server):   28.88ms
Avg prefill len:     2199.10 tokens
P50 prefill len:     2199.00 tokens
P99 prefill len:     2233.00 tokens
Avg decode len:      150.00 tokens
P50 decode len:      150.00 tokens
P99 decode len:      150.00 tokens
```

Also confirmed from GPU trace that there is no blocking behavior between vLLM model forward CUDA stream with other CUDA streams.
<img width="2335" height="881" alt="Screenshot 2025-07-30 at 10 19 28 PM" src="https://github.com/user-attachments/assets/e25ac78a-87bd-44a5-aca7-38dc839b6711" />
<img width="2337" height="1190" alt="Screenshot 2025-07-30 at 10 20 23 PM" src="https://github.com/user-attachments/assets/5b9a61c2-2c4e-4348-88c1-d78fc9f0fca6" />

## Accuracy
gsm8K 8shot (disagg)
```
[2025-07-30 23:06:36,254] [rank 0] [INFO] Per prompt detailed info dumped to /tmp/eval_dump.gsm8k.8_shot.1_gen.20250730_230636.json
[2025-07-30 23:06:36,254] [rank 0] [INFO] Evaluation results on task gsm8k.8_shot.1_gen: em: 0.970000 | f1: 0.970000 | em_maj1@1: 0.970000 | f1_maj1@1: 0.970000
[2025-07-30 23:06:36,254] [rank 0] [INFO] Task gsm8k.8_shot.1_gen took 54.20 seconds
```

mmlu_pro (non-disagg)
```
vllm (pretrained=/data/users/zijingliu/cp/Llama-4-Maverick-17B-128E-Instruct-FP8,dtype=auto,max_model_len=8196,tensor_parallel_size=8,gpu_memory_utilization=0.9,max_gen_toks=2048,seed=0), gen_kwargs: (None), limit: None, num_fewshot: None, batch_size: auto
|       Tasks       |Version|    Filter    |n-shot|  Metric   |   |Value |   |Stderr|
|-------------------|------:|--------------|-----:|-----------|---|-----:|---|-----:|
|mmlu_pro           |    2.0|custom-extract|      |exact_match|↑  |0.7976|±  |0.0036|
| - biology         |    2.1|custom-extract|     5|exact_match|↑  |0.8982|±  |0.0113|
| - business        |    2.1|custom-extract|     5|exact_match|↑  |0.8276|±  |0.0135|
| - chemistry       |    2.1|custom-extract|     5|exact_match|↑  |0.8445|±  |0.0108|
| - computer_science|    2.1|custom-extract|     5|exact_match|↑  |0.8220|±  |0.0189|
| - economics       |    2.1|custom-extract|     5|exact_match|↑  |0.8578|±  |0.0120|
| - engineering     |    2.1|custom-extract|     5|exact_match|↑  |0.7131|±  |0.0145|
| - health          |    2.1|custom-extract|     5|exact_match|↑  |0.7763|±  |0.0146|
| - history         |    2.1|custom-extract|     5|exact_match|↑  |0.6929|±  |0.0237|
| - law             |    2.1|custom-extract|     5|exact_match|↑  |0.5967|±  |0.0148|
| - math            |    2.1|custom-extract|     5|exact_match|↑  |0.8949|±  |0.0083|
| - other           |    2.1|custom-extract|     5|exact_match|↑  |0.7662|±  |0.0139|
| - philosophy      |    2.1|custom-extract|     5|exact_match|↑  |0.7355|±  |0.0198|
| - physics         |    2.1|custom-extract|     5|exact_match|↑  |0.8514|±  |0.0099|
| - psychology      |    2.1|custom-extract|     5|exact_match|↑  |0.8095|±  |0.0139|

| Groups |Version|    Filter    |n-shot|  Metric   |   |Value |   |Stderr|
|--------|------:|--------------|------|-----------|---|-----:|---|-----:|
|mmlu_pro|      2|custom-extract|      |exact_match|↑  |0.7976|±  |0.0036|
```



cc @KuntaiDu @simon-mo @WoosukKwon @houseroad @KingsleyZhang123