@sdavidbd 2025-07-31
> > With gpu model runner runs in default CUDA stream, the blocking copy operations such like valid_sampled_token_ids = sampled_token_ids.tolist() would block other copy operations scheduled
> 
> noob question, what's so special about the main stream that will block ops on other streams?

@yinghai, The default stream synchronizes with all blocking streams (they cannot overlap), see: https://docs.nvidia.com/cuda/cuda-runtime-api/stream-sync-behavior.html.