@sdavidbd 2025-07-31
Thanks @liuzijing2014! Appreciate the effort to improve stream efficiency here - it's definitely important in disagg setups with concurrent GPU activity. From what I understand, the issue arises during async KV load, where custom loading streams are blocked by the default stream used by the model runner.

That said, it seems this could be addressed more simply by creating the KV loading streams with the `cudaStreamNonBlocking` flag, which allows them to proceed without being blocked by the default stream.

Also, switching the model runner to a non-default stream might have unintended side effects on existing connector implementations (including ours) that must know which stream the model runner uses in order to correctly synchronize KV layer-wise (sync-)load/save operations with attention computation.