@liuzijing2014 2025-08-01
> Collaborator

@houseroad Tested with a larger volume (5120 requests), i do notice there is a slight regression: QPS (12.94 -> 12.81), TTFT (201.37 -> 206.31), TTIT(32.29 -> 32.64). Will try out on other OSS models QWEN/DeepSeek to see if this is widely observed.