@liuzijing2014 2025-08-05
@youkaichao @sdavidbd I added a new env var to control whether to use default stream or a separate stream.

For the suggestion around why not using a `non-blocking` CUDA stream within KV connector, we have tried that but the observation is that it wouldn't resolve the blocking issue across multiple streams. We made a post in PyTorch forum for this issue here: https://discuss.pytorch.org/t/default-pytorch-cuda-stream-copy-ops-surprisingly-block-other-copy-ops-from-other-pytorch-non-blocking-cuda-stream/221991

Besides, you can try the reproduce script above, which would show a similar issue.